[ARM][AArch64][TTI] Fix getArithmeticReductionCost() for non-FPMathOp w/o FMF.
In recent VPlan recipes chagnes, VPReductionRecipe is no longer contains
fast-math-flags for non-FPMathOperators. And according to the LangRef,
the fast-math flags only supported by thre return type is FP or vector
FP (for reduction).

This patch fix the cost of the reduction when the opcode is *not*
FPMathOperator and query the cost *without* FMF.